# openapi.json
{"openapi":"3.1.0","info":{"title":"Blog Index Analyzer","description":"네이버 블로그 지수 측정 및 분석 API","version":"v1"},"paths":{"/":{"get":{"summary":"Root","description":"헬스 체크","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"상세 헬스 체크","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user.\n\nCreates a new user account and returns an access token.","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login with email and password.\n\nReturns an access token if credentials are valid.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","description":"Get current user information.\n\nRequires authentication.","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["Authentication"],"summary":"Update Me","description":"Update current user information.\n\nRequires authentication.","operationId":"update_me_api_auth_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"delete":{"tags":["Authentication"],"summary":"Delete Me","description":"Delete current user account (soft delete).\n\nRequires authentication.","operationId":"delete_me_api_auth_me_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"HTTPBearer":[]}]}},"/api/blogs/analyze":{"post":{"tags":["블로그 분석"],"summary":"Analyze Blog","description":"블로그 분석 요청 (동기 실행)\n\n- **blog_id**: 네이버 블로그 ID\n- **analysis_type**: 'quick' (빠른 분석), 'full' (전체 분석), 또는 'manual' (수동 입력)\n- **manual_stats**: analysis_type이 'manual'인 경우 사용자가 직접 입력한 통계","operationId":"analyze_blog_api_blogs_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogAnalysisResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blogs/{blog_id}/index":{"get":{"tags":["블로그 분석"],"summary":"Get Blog Index","description":"블로그 지수 조회\n\n가장 최근 분석 결과를 반환합니다.\n\n- **blog_id**: 네이버 블로그 ID","operationId":"get_blog_index_api_blogs__blog_id__index_get","parameters":[{"name":"blog_id","in":"path","required":true,"schema":{"type":"string","title":"Blog Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blogs/{blog_id}/exists":{"get":{"tags":["블로그 분석"],"summary":"Check Blog Exists","description":"블로그 존재 여부 및 분석 이력 확인\n\n- **blog_id**: 네이버 블로그 ID","operationId":"check_blog_exists_api_blogs__blog_id__exists_get","parameters":[{"name":"blog_id","in":"path","required":true,"schema":{"type":"string","title":"Blog Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blogs/search-keyword":{"post":{"tags":["블로그 분석"],"summary":"Search Keyword And Analyze","description":"키워드로 블로그 검색 및 상위 N개 분석\n\n- **keyword**: 검색할 키워드\n- **limit**: 분석할 블로그 개수 (기본 100개, 최대 100개)","operationId":"search_keyword_and_analyze_api_blogs_search_keyword_post","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string","title":"Keyword"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/blogs/test/hello":{"get":{"tags":["블로그 분석"],"summary":"Test Endpoint","description":"테스트 엔드포인트","operationId":"test_endpoint_api_blogs_test_hello_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BlogAnalysisRequest":{"properties":{"blog_id":{"type":"string","maxLength":50,"minLength":1,"title":"Blog Id","description":"네이버 블로그 ID"},"analysis_type":{"type":"string","pattern":"^(quick|full|manual)$","title":"Analysis Type","description":"분석 타입","default":"quick"},"post_limit":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Post Limit","description":"분석할 포스트 수","default":10},"manual_stats":{"anyOf":[{"$ref":"#/components/schemas/ManualStats"},{"type":"null"}],"description":"수동 입력 통계 (analysis_type이 manual인 경우 사용)"}},"type":"object","required":["blog_id"],"title":"BlogAnalysisRequest","description":"블로그 분석 요청"},"BlogAnalysisResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"},"estimated_time_seconds":{"type":"integer","title":"Estimated Time Seconds"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result"}},"type":"object","required":["job_id","status","message","estimated_time_seconds"],"title":"BlogAnalysisResponse","description":"블로그 분석 응답"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ManualStats":{"properties":{"total_posts":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Posts"},"total_visitors":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Total Visitors"},"neighbor_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Neighbor Count"}},"type":"object","title":"ManualStats","description":"수동 입력 통계"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"Token","description":"Token response"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"name":{"type":"string","title":"Name"},"password":{"type":"string","maxLength":100,"minLength":8,"title":"Password"}},"type":"object","required":["email","name","password"],"title":"UserCreate","description":"User creation schema"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"User login schema"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"name":{"type":"string","title":"Name"},"is_active":{"type":"boolean","title":"Is Active"},"is_verified":{"type":"boolean","title":"Is Verified"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","name","is_active","is_verified","created_at"],"title":"UserResponse","description":"User response schema (without password)"},"UserUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},"type":"object","title":"UserUpdate","description":"User update schema"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}